small refactorings for to_cpp1.h file
